Issue #24: Support for consolidation/robo:5.x
