+++ .github/workflows/ci.yml
@@ -1,18 +1,24 @@
-name: CI
+name: 'CI'
 
-on: [push, pull_request_target]
+on:
+  - 'push'
+  - 'pull_request_target'
 
 jobs:
   build-test:
-    runs-on: ubuntu-latest
+    runs-on: 'ubuntu-latest'
     strategy:
       matrix:
-        PHP_VERSION: [ "8.0", "8.1" ]
+        php_version:
+          - '8.3'
 
     steps:
-      - uses: actions/checkout@v2
-      - uses: php-actions/composer@v6
+      -
+        uses: 'actions/checkout@v2'
+      -
+        uses: 'php-actions/composer@v6'
         with:
-          php_version: ${{ matrix.php_version }}
           version: 2
-      - uses: php-actions/phpunit@v3
+          php_version: ${{ matrix.php_version }}
+      -
+        uses: 'php-actions/phpunit@v3'
