+++ grumphp.yml.dist
@@ -7,7 +7,7 @@ grumphp:
       standard: vendor/drupal/coder/coder_sniffer/Drupal/
       ignore_patterns:
         - vendor/
-        - tests/
+        - tests/fixtures/
       triggered_by:
         - php
     git_commit_message:
