+++ phpcs.xml.dist
@@ -1,0 +1,22 @@
+<?xml version="1.0"?>
+<ruleset
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:noNamespaceSchemaLocation="https://raw.githubusercontent.com/squizlabs/PHP_CodeSniffer/master/phpcs.xsd"
+  name="Custom">
+
+  <file>./src/</file>
+  <file>./tests/src/</file>
+
+  <rule ref="Drupal">
+    <exclude name="Drupal.Commenting.ClassComment.Missing" />
+    <exclude name="Drupal.Commenting.Deprecated" />
+    <exclude name="Drupal.Commenting.DocComment.MissingShort" />
+    <exclude name="Drupal.Commenting.FunctionComment.Missing" />
+    <exclude name="Drupal.Commenting.FunctionComment.MissingParamComment" />
+    <exclude name="Drupal.Commenting.FunctionComment.MissingReturnComment" />
+    <exclude name="Drupal.Commenting.InlineComment.DocBlock" />
+    <exclude name="Drupal.Commenting.VariableComment.Missing" />
+    <exclude name="Drupal.Commenting.VariableComment.MissingVar" />
+  </rule>
+
+</ruleset>
